Model Rendering with Faces
In this push I added back in the model rendering but now we are rendering the faces of the models and not just the wireframe version of the model.

I also added in a simple trick for back face culling and triangles that cover less than one pixel but it is not perfect.